:bug: Kill prev stream on new job
We should kill the precious stream so that we don't listen to unwanted replies whenever there is a new daily prompt
