+++ util/helpers.py
@@ -1,4 +1,5 @@
 import sys
+from typing import List
 
 import pytz
 import tweepy
@@ -6,7 +7,7 @@ from loguru import logger
 from tweepy import Tweet
 from tweepy.models import Status
 
-from api.playlistter_bot import PlaylistterBot
+# from api.playlistter_bot import PlaylistterBot
 
 # Temporary In-memory map to keep track of user replies
 # Will get cleared each new day
@@ -37,14 +38,10 @@ def clear_user_reply_map() -> bool:
         return False
 
 
-def populate_user_replies_map(playlistter: PlaylistterBot, last_tweet: Status):
+def populate_user_replies_map(replies: List[Status]):
     global USER_REPLIES
-
-    # Since there is no way to directly grab the replies to a tweet, we need to use the search API
-    for tweet in tweepy.Cursor(playlistter.twitter_client.search_tweets, q=f"to:{last_tweet.author.screen_name}", result_type="recent") \
-            .items(500):
-        if hasattr(tweet, "in_reply_to_status_id_str") and tweet.in_reply_to_status_id_str == last_tweet.id_str:
-            if tweet.author.id_str not in USER_REPLIES:
-                song = tweet.text.strip()
-                logger.debug(f"Previous song suggestion by {tweet.author.id_str}: {tweet.text}")
-                USER_REPLIES[tweet.author.id_str] = song
+    for reply in replies:
+        if reply.author.id_str not in USER_REPLIES:
+            song = reply.text.strip()
+            logger.debug(f"Previous song suggestion by {reply.author.id_str}: {reply.text}")
+            USER_REPLIES[reply.author.id_str] = song
